requirements.txt
python-dotenv
openai 
psycopg
pytest